+++ src/main/java/com/vbox/persistent/pojo/param/DeptParam.java
@@ -7,10 +7,10 @@ import java.time.LocalDateTime;
 @Data
 public class DeptParam {
 
-    private Long id;
-    private Long pid;
-    private Long parentDept;
-    private Long orderNo;
+    private Integer id;
+    private Integer pid;
+    private Integer parentDept;
+    private Integer orderNo;
     private String deptName;
     private String status;
     private String remark;
