+++ src/main/java/com/vbox/persistent/pojo/param/UserCreateOrUpdParam.java
@@ -7,9 +7,9 @@ import java.time.LocalDateTime;
 @Data
 public class UserCreateOrUpdParam {
 
-    private Long id;
-    private Long roleId;
-    private Long deptId;
+    private Integer id;
+    private Integer roleId;
+    private Integer deptId;
     private String account;
     private Integer gender;
     private String pass;
