+++ src/main/java/com/vbox/persistent/pojo/vo/DeptVO.java
@@ -8,10 +8,10 @@ import java.util.List;
 @Data
 public class DeptVO {
 
-    private Long id;
-    private Long pid;
-    private Long parentDept;
-    private Long orderNo;
+    private Integer id;
+    private Integer pid;
+    private Integer parentDept;
+    private Integer orderNo;
     private String deptName;
     private String status;
     private String remark;
