+++ src/main/java/com/vbox/persistent/pojo/vo/RoleMenuVO.java
@@ -8,12 +8,12 @@ import java.util.List;
 @Data
 public class RoleMenuVO {
 
-    private Long id;
-    private Long orderNo;
+    private Integer id;
+    private Integer orderNo;
     private String roleName;
     private String roleValue;
     private LocalDateTime createTime;
-    private List<Long> menu;
+    private List<Integer> menu;
     private String remark;
     private String status;
 }
