+++ src/main/java/com/vbox/persistent/pojo/vo/RoleVO.java
@@ -9,8 +9,8 @@ import java.util.List;
 @Data
 public class RoleVO {
 
-    private Long id;
-    private Long orderNo;
+    private Integer id;
+    private Integer orderNo;
     private String roleName;
     private String roleValue;
     private LocalDateTime createTime;
